LIME_TriggerEventsFromSceneOffsetIfFollowing takes eleven, not four
The frame says so: `push {r4,r5,r6,r7,lr}` (20) + `push.w {r8,sl,fp}` (12) +
`sub sp, #0x68` (104) = 136 = 0x88, so `[sp, #0x88]` is the caller's first
stacked argument -- and the body reads every word from there to `[sp, #0xa0]`,
which is seven more. GameCode.c had always declared eleven, with names for all
of them. The body still models only the first four; the rest are declared so
the interface is right and marked unused so it is obvious which half is read.

Three spellings in parkfn, each of which had cost it between forty and eighty
functions: the dispatch is `cbnz rTok, RESUME` or `cmp rTok, #0` with `bne`,
and the token check is `cmp` or `cmp.w`. It also takes a store whose value
came from another field now, with the same saved-value guard leaffn needed.

When a reader's yield is surprisingly low, suspect the spelling before the
logic. That is three times today.

Cross-file prototypes: 3 left, and one of those is the float-ABI question
measured in #26.

+++ decomp/gamecode/Blood.c
@@ -626,7 +626,7 @@ void  limeMatrixCopy(const float *src, float *dst);
 void  limeScaleMatrix(float *mtx, float s);
 void  RotMatrixX(float *mtx, float angle);
 void  LIME_PlayFBXAtPos(float *mtx, long a, void *scene, long b);
-void  LIME_TriggerEventsFromScene(void *scene, long frame, const float *mtx,
+void  LIME_TriggerEventsFromScene(void *scene, int frame, const float *mtx,
                                   long flags, long a, long b, long c, long d);
 long  isParent(void);
 long  isParentBasedOnSpeed(void);
